+++ CHANGELOG.md
@@ -1,3 +1,10 @@
 ## 1.0.0
 
 - Initial version.
+
+
+## 1.0.1
+
+- Bit more DOC information
+- Terminal class for the output to the terminal
+- Cleaned up the constructors